+++ lib/views/home_content/challenge/challenge_list.dart
@@ -1,7 +1,7 @@
 import "package:flutter/material.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
-import "package:proxima/utils/ui/circular_value.dart";
 import "package:proxima/viewmodels/challenge_view_model.dart";
+import "package:proxima/views/components/async/circular_value.dart";
 import "package:proxima/views/home_content/challenge/challenge_card.dart";
 
 class ChallengeList extends ConsumerWidget {
