+++ lib/views/home_content/map/map_screen.dart
@@ -1,7 +1,7 @@
 import "package:flutter/material.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
-import "package:proxima/utils/ui/circular_value.dart";
 import "package:proxima/viewmodels/map_view_model.dart";
+import "package:proxima/views/components/async/circular_value.dart";
 import "package:proxima/views/home_content/map/post_map.dart";
 import "package:proxima/views/option_widgets/map/map_selection_option_chips.dart";
