+++ lib/views/pages/new_post/new_post_form.dart
@@ -1,8 +1,8 @@
 import "package:flutter/material.dart";
 import "package:flutter_hooks/flutter_hooks.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
-import "package:proxima/utils/ui/circular_value.dart";
 import "package:proxima/viewmodels/new_post_view_model.dart";
+import "package:proxima/views/components/async/circular_value.dart";
 
 class NewPostForm extends HookConsumerWidget {
   const NewPostForm({super.key});
