+++ test/mocks/providers/provider_circular_value.dart
@@ -1,6 +1,6 @@
 import "package:flutter/material.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
-import "package:proxima/utils/ui/circular_value.dart";
+import "package:proxima/views/components/async/circular_value.dart";
 
 ProviderScope circularValueProvider(AsyncValue<void> value) {
   return ProviderScope(
