+++ test/viewmodels/challenge_view_model_integration_test.dart
@@ -6,9 +6,9 @@ import "package:proxima/services/database/challenge_repository_service.dart";
 import "package:proxima/services/database/firestore_service.dart";
 import "package:proxima/services/database/user_repository_service.dart";
 import "package:proxima/services/geolocation_service.dart";
-import "package:proxima/utils/ui/circular_value.dart";
 import "package:proxima/viewmodels/challenge_view_model.dart";
 import "package:proxima/viewmodels/login_view_model.dart";
+import "package:proxima/views/components/async/circular_value.dart";
 
 import "../mocks/data/firestore_challenge.dart";
 import "../mocks/data/firestore_post.dart";
